refactor(rsc): remove extra 'use client' and split heavy widgets
