+++ src/features/reviews/ui/TestimonialSliderWrapper.tsx
@@ -1,6 +1,37 @@
-'use client';
+import React from 'react';
+import dynamic from 'next/dynamic';
+import { Testimonial } from '../';
 
-import { TestimonialSlider, Testimonial } from '../';
+// Динамический импорт TestimonialSlider с загрузочным скелетоном
+const TestimonialSlider = dynamic(() => import('./TestimonialSlider'), {
+  ssr: false,
+  loading: () => (
+    <div className="mx-auto mt-0 flex max-w-2xl flex-col items-center">
+      <div className="relative h-auto min-h-[380px] w-full max-w-2xl rounded-xl border border-[#FFFFFF25] bg-gradient-to-br from-[#1A1A1A] to-[#2A2A2A] p-6 shadow-2xl md:p-8">
+        <div className="flex h-full flex-col text-center">
+          <div className="flex flex-1 flex-col justify-center">
+            <div className="mb-2 h-6 w-32 animate-pulse rounded bg-gray-700 mx-auto"></div>
+            <div className="mb-4 h-4 w-24 animate-pulse rounded bg-gray-600 mx-auto"></div>
+            <div className="space-y-2">
+              <div className="h-4 w-full animate-pulse rounded bg-gray-600"></div>
+              <div className="h-4 w-5/6 animate-pulse rounded bg-gray-600 mx-auto"></div>
+              <div className="h-4 w-4/6 animate-pulse rounded bg-gray-600 mx-auto"></div>
+            </div>
+          </div>
+        </div>
+      </div>
+      <div className="mt-6 flex items-center gap-4">
+        <div className="h-9 w-9 animate-pulse rounded-full bg-gray-700"></div>
+        <div className="flex gap-2">
+          {[...Array(3)].map((_, i) => (
+            <div key={i} className="h-2 w-2 animate-pulse rounded-full bg-gray-700"></div>
+          ))}
+        </div>
+        <div className="h-9 w-9 animate-pulse rounded-full bg-gray-700"></div>
+      </div>
+    </div>
+  ),
+});
 
 const TestimonialSliderWrapper = () => {
   // Используем те же данные отзывов, что и в оригинальном компоненте
